0 indexed, other stuff i forgot

+++ src/model/ohlc_candles.rs
@@ -16,7 +16,7 @@ impl OHLCCandles {
 
 impl RendererExtension for OHLCCandles {
 	fn apply(&self, buffer: &mut ChartBuffer, data: &[OHLC]) {
-		let period = buffer.timeframe / data.len();
+		let period = buffer.timeframe / data.len() as i64;
 
 		for i in 0..data.len() {
 			let ohlc = data[i];
@@ -24,14 +24,14 @@ impl RendererExtension for OHLCCandles {
 			let colour = if ohlc.o > ohlc.c { self.down_colour } else { self.up_colour };
 
 			{
-				let p1 = buffer.data_to_coords(ohlc.o, period * i);
-				let p2 = buffer.data_to_coords(ohlc.c, period * (i + 1) - 1);
+				let p1 = buffer.data_to_coords(ohlc.o, period * i as i64);
+				let p2 = buffer.data_to_coords(ohlc.c, period * (i as i64 + 1) - 1);
 
 				buffer.rect_point(p1, p2, colour);
 			}
 
 			{
-				let time = period * i + (period / 2);
+				let time = period * i as i64 + (period / 2);
 				let p1 = buffer.data_to_coords(ohlc.h, time - (period / 8));
 				let p2 = buffer.data_to_coords(ohlc.l, time + (period / 8));
